finding email from security.txt + fixed crashing when e-mail was in javascript
